+++ android/common/NioUtils.h
