+++ android/filament-android/src/main/cpp/IndexBuffer.cpp
@@ -22,8 +22,8 @@
 
 #include <filament/IndexBuffer.h>
 
-#include "CallbackUtils.h"
-#include "NioUtils.h"
+#include "common/CallbackUtils.h"
+#include "common/NioUtils.h"
 
 using namespace filament;
 using namespace backend;
